adds additional check for isRestricedNode

+++ boot.sh
@@ -17,12 +17,13 @@ fi
 
 cp -r "$IMAGE_REQEXEC_DIR"/. "$REQEXEC_DIR"
 
-if [ -n "$KEY_STORE_LOCATION" ]; then
-  rm -rf $KEY_STORE_LOCATION/*
-fi
-
-if [ -n "$MESSAGE_STORE_LOCATION" ]; then
-  rm -rf $MESSAGE_STORE_LOCATION/*
+if [ "$IS_RESTRICTED_NODE" == "true" ]; then
+  if [ -n "$KEY_STORE_LOCATION" ]; then
+    rm -rf $KEY_STORE_LOCATION/*
+  fi
+  if [ -n "$MESSAGE_STORE_LOCATION" ]; then
+    rm -rf $MESSAGE_STORE_LOCATION/*
+  fi
 fi
 
 if [ "$RUN_MODE" == "dev" ]; then
